Update GameView.java
The game now prompts the user to return to main menu or exit once all levels are complete.

If the user returns to main menu and clicks start again, he will be put in the first level.

This will most likely change once level builder is created / the form where the user picks what level they want to start at.